refactor(chat): split quick chat rendering
